Move MergeLuaAst trait and impls into own file
